# marketstack-openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Marketstack OpenAPI",
    "description": "Inofficial marketstack OpenAPI spec",
    "version": "1.0"
  },
  "paths": {
...
    "/intraday": {
      "get": {
        "tags": [
          "intraday"
        ],
        "summary": "Query",
        "operationId": "intraday",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          },
          {
            "required": true,
            "schema": {
              "title": "Comma-separated symbols list",
              "type": "string"
            },
            "example": "AAPL,AMZN",
            "name": "symbols",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Exchange MIC",
              "type": "string"
            },
            "example": "XNAS",
            "name": "exchange",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Date/time sort order",
              "allOf": [
                {
                  "$ref": "#/components/schemas/Sort"
                }
              ]
            },
            "example": "DESC",
            "name": "sort",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Intraday interval",
              "allOf": [
                {
                  "$ref": "#/components/schemas/Interval"
                }
              ]
            },
            "example": "5min",
            "name": "interval",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_from",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_to",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "maximum": 1000.0,
              "minimum": 1.0,
              "type": "integer"
            },
            "example": 10,
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0.0,
              "type": "integer"
            },
            "example": 0,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_List_models.IntervalPrice__"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/tickers": {
      "get": {
        "tags": [
          "tickers"
        ],
        "summary": "Query",
        "operationId": "tickers",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Exchange MIC",
              "type": "string"
            },
            "example": "XNAS",
            "name": "exchange",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Search string",
              "type": "string"
            },
            "name": "search",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "maximum": 1000.0,
              "minimum": 1.0,
              "type": "integer"
            },
            "example": 10,
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0.0,
              "type": "integer"
            },
            "example": 0,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_List_models.Ticker__"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/exchanges": {
      "get": {
        "tags": [
          "exchanges"
        ],
        "summary": "Query",
        "operationId": "exchanges",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Search string",
              "type": "string"
            },
            "name": "search",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "maximum": 1000.0,
              "minimum": 1.0,
              "type": "integer"
            },
            "example": 10,
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0.0,
              "type": "integer"
            },
            "example": 0,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_List_models.Exchange__"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/exchanges/{mic}": {
      "get": {
        "tags": [
          "exchanges"
        ],
        "summary": "Mic",
        "operationId": "exchange_mic",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Exchange MIC",
              "type": "string"
            },
            "example": "XNAS",
            "name": "mic",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exchange"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/exchanges/{mic}/tickers": {
      "get": {
        "tags": [
          "exchanges"
        ],
        "summary": "Tickers",
        "operationId": "exchange_mic_tickers",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Exchange MIC",
              "type": "string"
            },
            "example": "XNAS",
            "name": "mic",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_ExchangeTickers_"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/exchanges/{mic}/eod": {
      "get": {
        "tags": [
          "exchanges"
        ],
        "summary": "Mic Eod",
        "operationId": "exchange_mic_eod",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Exchange MIC",
              "type": "string"
            },
            "example": "XNAS",
            "name": "mic",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          },
          {
            "required": true,
            "schema": {
              "title": "Comma-separated symbols list",
              "type": "string"
            },
            "example": "AAPL,AMZN",
            "name": "symbols",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Date/time sort order",
              "allOf": [
                {
                  "$ref": "#/components/schemas/Sort"
                }
              ]
            },
            "example": "DESC",
            "name": "sort",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_from",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_to",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "maximum": 1000.0,
              "minimum": 1.0,
              "type": "integer"
            },
            "example": 10,
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0.0,
              "type": "integer"
            },
            "example": 0,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_ExchangeEod_"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/exchanges/{mic}/eod/latest": {
      "get": {
        "tags": [
          "exchanges"
        ],
        "summary": "Mic Eod Latest",
        "operationId": "exchange_mic_eod_latest",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Exchange MIC",
              "type": "string"
            },
            "example": "XNAS",
            "name": "mic",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          },
          {
            "required": true,
            "schema": {
              "title": "Comma-separated symbols list",
              "type": "string"
            },
            "example": "AAPL,AMZN",
            "name": "symbols",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_ExchangeEod_"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/exchanges/{mic}/eod/{date}": {
      "get": {
        "tags": [
          "exchanges"
        ],
        "summary": "Mic Eod Date",
        "operationId": "exchange_mic_eod_date",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Exchange MIC",
              "type": "string"
            },
            "example": "XNAS",
            "name": "mic",
            "in": "path"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": true,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Comma-separated symbols list",
              "type": "string"
            },
            "example": "AAPL,AMZN",
            "name": "symbols",
            "in": "query"
          },
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_ExchangeEod_"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/exchanges/{mic}/intraday": {
      "get": {
        "tags": [
          "exchanges"
        ],
        "summary": "Mic Intraday",
        "operationId": "exchange_mic_intraday",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Exchange MIC",
              "type": "string"
            },
            "example": "XNAS",
            "name": "mic",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          },
          {
            "required": true,
            "schema": {
              "title": "Comma-separated symbols list",
              "type": "string"
            },
            "example": "AAPL,AMZN",
            "name": "symbols",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Intraday interval",
              "allOf": [
                {
                  "$ref": "#/components/schemas/Interval"
                }
              ]
            },
            "example": "5min",
            "name": "interval",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Date/time sort order",
              "allOf": [
                {
                  "$ref": "#/components/schemas/Sort"
                }
              ]
            },
            "example": "DESC",
            "name": "sort",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_from",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_to",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "maximum": 1000.0,
              "minimum": 1.0,
              "type": "integer"
            },
            "example": 10,
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0.0,
              "type": "integer"
            },
            "example": 0,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_ExchangeIntraday_"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "title": "ErrorResponse",
        "required": [
          "error"
        ],
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          }
        }
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "Interval": {
        "title": "Interval",
        "enum": [
          "1min",
          "5min",
          "10min",
          "30min",
          "1hour",
          "3hour",
          "6hour",
          "12hour",
          "24hour"
        ],
        "description": "An enumeration."
      },
...
      "PagedResponse_ExchangeEod_": {
        "title": "PagedResponse[ExchangeEod]",
        "required": [
          "pagination",
          "data"
        ],
        "type": "object",
        "properties": {
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          },
          "data": {
            "$ref": "#/components/schemas/ExchangeEod"
          }
        }
      },
...
      "PagedResponse_ExchangeTickers_": {
        "title": "PagedResponse[ExchangeTickers]",
        "required": [
          "pagination",
          "data"
        ],
        "type": "object",
        "properties": {
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          },
          "data": {
            "$ref": "#/components/schemas/ExchangeTickers"
          }
        }
      },
...
      "PagedResponse_List_models.Exchange__": {
        "title": "PagedResponse[List[models.Exchange]]",
        "required": [
          "pagination",
          "data"
        ],
        "type": "object",
        "properties": {
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          },
          "data": {
            "title": "Data",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Exchange"
            }
          }
        }
      },
      "PagedResponse_List_models.IntervalPrice__": {
        "title": "PagedResponse[List[models.IntervalPrice]]",
        "required": [
          "pagination",
          "data"
        ],
        "type": "object",
        "properties": {
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          },
          "data": {
            "title": "Data",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IntervalPrice"
            }
          }
        }
      },
...
      "PagedResponse_List_models.Ticker__": {
        "title": "PagedResponse[List[models.Ticker]]",
        "required": [
          "pagination",
          "data"
        ],
        "type": "object",
        "properties": {
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          },
          "data": {
            "title": "Data",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Ticker"
            }
          }
        }
      },
...
      "Sort": {
        "title": "Sort",
        "enum": [
          "ASC",
          "DESC"
        ],
        "description": "An enumeration."
      },
...
    }
  }
}